Refactor SDK node module into subpackages
